Fix- stop app from listening and refactore
